+++ src/pydantic_mermaid/mermaid_generator.py
@@ -1,85 +1,96 @@
+from copy import deepcopy
 from types import ModuleType
 from typing import Set, Type, Dict
 
-from pydantic_mermaid.models import MermaidGraph, Relations
+from importlib_resources import files
+from jinja2 import Environment, FileSystemLoader
+
+from pydantic_mermaid.models import MermaidClass, MermaidGraph, Relations
 from pydantic_mermaid.pydantic_parser import PydanticParser
+
+Map = List[Tuple[str, str]]
+
+
+def render(classes: List[MermaidClass], client_services: Map, parent_children: Map) -> str:
+    """Render a mermaid graph to a string"""
+
+    template_path = files("pydantic_mermaid")
+    env = Environment(loader=FileSystemLoader(str(template_path)), trim_blocks=True, lstrip_blocks=True)
+    template = env.get_template("template.j2")
+    return template.render(classes=classes, client_services=client_services, parent_children=parent_children)
 
 
 class MermaidGenerator:
     """genertate a class chart from module"""
 
-    def __init__(self, module: ModuleType, *, ignore_type_dependencies: list[Type] = (dict, list, set)) -> None:
-        self.g: MermaidGraph = PydanticParser()(module, ignore_type_dependencies)
+    def __init__(self, module: ModuleType) -> None:
+        self.graph: MermaidGraph = PydanticParser()(module)
         self.allow_set: Set[str] = set()
 
     def generate_allow_list(self, root: str, relations: Relations) -> None:
         """
-        user can focus on certain class `root` and prune classes that is inherited from `root`
-        or not a dependencies of `root`
+        user can focus on certain class `root` and prune classes that is not inherited from `root`
+        or does not depend on `root`
         """
         self.allow_set = {root}
-        reversed_class_names = reversed(self.g.class_names)
+        reversed_class_names = reversed(self.graph.class_names)
         if root != "" and relations & Relations.Dependency:
             for parent in reversed_class_names:
-                if parent in self.allow_set and parent in self.g.service_clients:
-                    self.allow_set = self.allow_set | self.g.service_clients[parent]
+                if parent in self.allow_set and parent in self.graph.service_clients:
+                    self.allow_set = self.allow_set | self.graph.service_clients[parent]
 
         if root != "" and relations & Relations.Inheritance:
             for parent in reversed_class_names:
-                if parent in self.allow_set and parent in self.g.parent_children:
-                    self.allow_set = self.allow_set | self.g.parent_children[parent]
+                if parent in self.allow_set and parent in self.graph.parent_children:
+                    self.allow_set = self.allow_set | self.graph.parent_children[parent]
 
         if root == "":
-            self.allow_set = set(self.g.class_dict)
+            self.allow_set = set(self.graph.class_dict)
 
-    def generate_dependencies(self) -> str:
-        """print dependencies for class chart"""
-        s = ""
-        for dependant, depended in self.g.service_clients.items():
-            if dependant not in self.allow_set:
+    def _generate_relationship_map(self, relationship: Dict[str, Set[str]]) -> Map:
+        """Generic method to generate relationship maps for dependencies or inheritance."""
+        relationship_map: Map = []
+        for entity, related_entities in relationship.items():
+            if entity not in self.allow_set:
                 continue
+            for related_entity in related_entities:
+                relationship_map.append((entity, related_entity))
+        return relationship_map
 
-            for d in depended:
-                s += f"    {dependant} ..> {d}\n"
+    def generate_dependencies(self) -> Map:
+        """Generate dependencies for class chart using the generic relationship method."""
+        return self._generate_relationship_map(self.graph.service_clients)
 
-        s += "\n"
-        return s
-
-    def generate_inheritance(self) -> str:
-        """print inheritance for class chart"""
-        s = ""
-        for parent, children in self.g.parent_children.items():
-            if parent not in self.allow_set:
-                continue
-            for child in children:
-                s += f"    {parent} <|-- {child}\n"
-        return s
+    def generate_inheritance(self) -> Map:
+        """Generate inheritance for class chart using the generic relationship method."""
+        return self._generate_relationship_map(self.graph.parent_children)
 
     def generate_chart(self, *, root: str = "", relations: Relations = Relations.Dependency) -> str:
         """print class chart"""
         self.generate_allow_list(root, relations)
 
-        s = "```mermaid\nclassDiagram"
-        for class_name, class_type in self.g.class_dict.items():
-            if class_name not in self.allow_set:
-                continue
+        final_classes: List[MermaidClass] = []
+
+        for class_name in self.allow_set:
+            class_value = self.graph.class_dict[class_name]
+
+            final_class = deepcopy(class_value)
 
             parent_class_name = ""
-            if class_name in self.g.child_parents:
-                parent_class_name = next(iter(self.g.child_parents[class_name]))
+            if class_name in self.graph.child_parents:
+                parent_class_name = next(iter(self.graph.child_parents[class_name]))
             if parent_class_name in self.allow_set and relations != Relations.Dependency:
-                inherited_properties = {f"{p.name}: {p.type}" for p in self.g.class_dict[parent_class_name].properties}
-                s += class_type.generate_class(exclude=inherited_properties)
-            else:
-                s += str(class_type)
+                inherited_properties = {str(p) for p in self.graph.class_dict[parent_class_name].properties}
+                final_class.properties = [p for p in final_class.properties if str(p) not in inherited_properties]
 
-        s += "\n\n"
+            final_classes.append(final_class)
 
+        client_services: Map = []
         if Relations.Dependency & relations:
-            s += self.generate_dependencies()
+            client_services = self.generate_dependencies()
 
+        parent_children: Map = []
         if Relations.Inheritance & relations:
-            s += self.generate_inheritance()
+            parent_children = self.generate_inheritance()
 
-        s += "```"
-        return s
+        return render(final_classes, client_services, parent_children)