+++ CHANGELOG.md
@@ -2,7 +2,19 @@
 
 This file documents all notable changes to this project.
 
-## [0.6.0] - TBD
+## [0.7.1] - 2024-02-25
+### Changed
+- Fix ellipsis in type
+
+## [0.7.0] - 2024-02-24
+
+### Changed
+- Support Enums
+- Add default values
+- Use jinja2 templates
+- Use strict mypy
+
+## [0.6.0] - 2024-01-14
 
 ### Changed
 - Support typing.Literal